Make hash take max bits and move MSB impl into ac_hsitory
model: gpt-5.6-luna
