+++ src/bin/eh-ctx-distribution.rs
@@ -23,12 +23,12 @@ fn main() -> io::Result<()> {
     inner.train(&buf);
     let inner = inner.freeze();
 
-    let mut history = ACHistory::new(7, inner);
+    let mut history = ACHistory::new(inner);
     let mut model = CtxPrefixModel8::new(Counter4::new());
     for byte in buf.iter() {
         unroll_for!(bit in byte, {
             history.update(bit);
-            model.set_ctx(history.hash());
+            model.set_ctx(history.hash(7));
             ctx_stats_ac[model.ctx] += 1;
         });
     }
