+++ src/history/ac_history.rs
@@ -3,7 +3,7 @@ use crate::{
     entropy_coding::arithmetic_coder::{ACWrite, ArithmeticCoder},
     helpers::RotatingBuffer,
     models::Model,
-    u8, Analytics,
+    u8,
 };
 
 #[derive(Clone)]
@@ -11,20 +11,18 @@ pub struct ACHistory<M: Model> {
     pos: usize,
     bits: u64,
     probs: RotatingBuffer<u16, 64>,
-    max_bits: u8,
-    compressed_bits_count: usize,
+    processed_bits: usize,
     model: M,
 }
 
 impl<M: Model> ACHistory<M> {
-    pub fn new(max_bits: u8, model: M) -> Self {
+    pub fn new(model: M) -> Self {
         const HALF: u16 = 1 << 15;
         Self {
             pos: 0,
             bits: 0,
             probs: RotatingBuffer::init(HALF),
-            max_bits,
-            compressed_bits_count: 0,
+            processed_bits: 0,
             model,
         }
     }
@@ -38,41 +36,105 @@ impl<M: Model> History for ACHistory<M> {
 
         self.bits = (self.bits << 1) | u64::from(bit);
         self.pos += 1;
-        self.compressed_bits_count = 0;
+        self.processed_bits = 0;
     }
 
-    fn hash(&mut self) -> u32 {
-        // if self.pos == 100 {
-        //     let mut entropy = 0.0;
-        //     for i in 0..64 {
-        //         let bit = u8!((self.bits >> i) & 1);
-        //         let p = self.cache[i];
-        //         let prob = f64::from(p) / 65536.0;
-        //         let prob = if bit == 1 { prob } else { 1.0 - prob };
-        //         entropy += -prob.log2();
-        //         println!("bit={}, p={}, h={}", bit, p, entropy);
-        //     }
-        // }
-
+    fn hash(&mut self, max_bits: u8) -> u32 {
         let mut ac = ArithmeticCoder::new_coder();
-        let mut writer = EntropyWriter {
-            state: 0,
-            rev_bits: 0,
-            idx: 0,
-            max_bits: self.max_bits,
-        };
+        let mut writer = EntropyWriter::new(max_bits);
         for i in 0..self.pos.min(64) {
             let bit = u8!((self.bits >> i) & 1);
             let p = self.probs[i];
-            let res = ac.encode(bit, p, &mut writer);
-            if res.is_err() {
-                self.compressed_bits_count = i + 1;
+            if ac.encode(bit, p, &mut writer).is_err() {
+                self.processed_bits = i + 1;
                 break;
             }
         }
-        // _ = ac.flush(&mut writer);
 
-        // writer.state >> (32 - writer.idx)
+        writer.state.wrapping_shr(32 - writer.idx as u32)
+    }
+}
+
+/// An entropy-coded history whose bits are arranged in MSB-first byte order.
+///
+/// `ACHistory` visits the history one bit at a time, starting with the most
+/// recently received bit. That makes the bits within every byte appear in
+/// reverse order. This variant starts with the partial (most recent) byte,
+/// but visits the bits in that byte in their original order. It then visits
+/// each preceding byte, also MSB first.
+#[derive(Clone)]
+pub struct ACHistoryMSB<M: Model> {
+    pos: usize,
+    bits: u64,
+    probs: RotatingBuffer<u16, 64>,
+    processed_bits: usize,
+    model: M,
+}
+
+impl<M: Model> ACHistoryMSB<M> {
+    pub fn new(model: M) -> Self {
+        const HALF: u16 = 1 << 15;
+        Self {
+            pos: 0,
+            bits: 0,
+            probs: RotatingBuffer::init(HALF),
+            processed_bits: 0,
+            model,
+        }
+    }
+}
+
+/// Return the index in the rotating history for the `ordinal`th bit in the
+/// MSB-first view of the history.
+///
+/// `partial_bits` is the number of bits in the current partial byte and is in
+/// the range `0..=7`. A value of zero means that the most recent byte is
+/// complete. The rotating buffer is indexed from the newest bit: index zero
+/// is the last bit received. The current partial byte and every preceding
+/// byte are therefore read in descending buffer-index order.
+#[inline]
+fn history_index_msb(ordinal: usize, partial_bits: usize) -> usize {
+    if ordinal < partial_bits {
+        return partial_bits - ordinal - 1;
+    }
+    let previous = ordinal - partial_bits;
+    let byte = previous / 8;
+    let bit = previous % 8;
+    partial_bits + byte * 8 + 7 - bit
+}
+
+impl<M: Model> History for ACHistoryMSB<M> {
+    fn update(&mut self, bit: u8) {
+        let p = self.model.predict();
+        self.model.update(bit);
+        self.probs.push(p);
+
+        self.bits = (self.bits << 1) | u64::from(bit);
+        self.pos += 1;
+        self.processed_bits = 0;
+    }
+
+    fn hash(&mut self, max_bits: u8) -> u32 {
+        let mut ac = ArithmeticCoder::new_coder();
+        let mut writer = EntropyWriter::new(max_bits);
+        let partial_bits = self.pos % 8;
+
+        for ordinal in 0..self.pos.min(64) {
+            let index = history_index_msb(ordinal, partial_bits);
+            // With an unaligned history, the next complete MSB-first byte
+            // can begin before the 64-bit raw window. Do not read beyond the
+            // rotating buffer just to fill the output to 64 bits.
+            if index >= 64 {
+                break;
+            }
+            let bit = u8!((self.bits >> index) & 1);
+            let p = self.probs[index];
+            if ac.encode(bit, p, &mut writer).is_err() {
+                self.processed_bits = ordinal + 1;
+                break;
+            }
+        }
+
         writer.state.wrapping_shr(32 - writer.idx as u32)
     }
 }
@@ -83,6 +145,12 @@ struct EntropyWriter {
     max_bits: u8,
     rev_bits: u16,
     idx: u8,
+}
+
+impl EntropyWriter {
+    fn new(max_bits: u8) -> Self {
+        Self { state: 0, max_bits, rev_bits: 0, idx: 0 }
+    }
 }
 
 impl ACWrite for EntropyWriter {
@@ -121,23 +189,36 @@ impl ACWrite for EntropyWriter {
     }
 }
 
-impl<M: Model + Analytics> Analytics for ACHistory<M> {
-    fn log(&mut self) -> serde_json::Value {
-        serde_json::json!({
-            "h": self.hash(),
-            "bits": self.compressed_bits_count,
-            "model": self.model.log(),
-        })
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn history_indices_are_msb_first_by_byte() {
+        // One bit in the current byte, followed by two complete preceding
+        // bytes: current, then 8..1, then 16..9.
+        let indices = (0..17)
+            .map(|i| super::history_index_msb(i, 1))
+            .collect::<Vec<_>>();
+        assert_eq!(
+            indices,
+            vec![0, 8, 7, 6, 5, 4, 3, 2, 1, 16, 15, 14, 13, 12, 11, 10, 9]
+        );
+
+        // Two bits in the current byte: current bits are 1,0, followed by
+        // the preceding byte 9..2.
+        let indices = (0..10)
+            .map(|i| super::history_index_msb(i, 2))
+            .collect::<Vec<_>>();
+        assert_eq!(indices, vec![1, 0, 9, 8, 7, 6, 5, 4, 3, 2]);
     }
 
-    fn metadata(&self) -> serde_json::Value {
-        serde_json::json!({
-            "type": "history/ACHistory",
-            "children": {
-                "model": self.model.metadata(),
-            },
-            "vars": { "bits": "usize", },
-            "description": "Compressed history with arithmetic coding using a static model",
-        })
+    #[test]
+    fn complete_current_byte_uses_zero_partial_bits() {
+        let indices = (0..17)
+            .map(|i| super::history_index_msb(i, 0))
+            .collect::<Vec<_>>();
+        assert_eq!(
+            indices,
+            vec![7, 6, 5, 4, 3, 2, 1, 0, 15, 14, 13, 12, 11, 10, 9, 8, 23]
+        );
     }
 }
