+++ src/history/mod.rs
@@ -1,12 +1,11 @@
 pub mod ac_history;
-pub mod ac_history_msb;
 pub mod huff_history;
 
-pub use self::{ac_history::*, ac_history_msb::*, huff_history::*};
+pub use self::{ac_history::*, huff_history::*};
 
 pub trait History {
     fn update(&mut self, bit: u8);
-    fn hash(&mut self) -> u32;
+    fn hash(&mut self, max_bits: u8) -> u32;
 }
 
 impl History for u32 {
@@ -14,7 +13,7 @@ impl History for u32 {
         *self = (*self << 1) | u32::from(bit);
     }
 
-    fn hash(&mut self) -> u32 {
+    fn hash(&mut self, _max_bits: u8) -> u32 {
         *self
     }
 }
